fix: address WIP pull review feedback
